Put direct keybindings in the minor mode map
Previously we used emulation-mode-map-alists because it provides a
straightforward way of toggling the keymaps based on conditions.  The
problem is, this variable puts the bindings at higher priority than
any other mode's bindings; it is undesirable to claim such a high
priority.

Instead, we rely on yas-expand-from-keymap to check conditions.